requirements.txt
PyVCF==0.6.8
mysql-connector-python==8.0.20
numpy==1.18.5
pandas==1.0.4
pip==20.1.1
protobuf==3.12.2
pythondateutil
pytz==2020.1
setuptools=46.4.0
six==1.15.0
flask==1.1.2
Werkzeug~=1.0.0
scikit-allel==1.2.1
frameup==0.1.2
hail~=0.2.41
